pacients rbac

+++ controllers/user.controller.js
@@ -1,60 +1,88 @@
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
+const decode = require("jwt-decode");
 const config = require("../config/config.json");
 const User = require("../models/user.model.js").User;
 const db = require("../config/database");
-const rules = require("../config/rules");
+const roles = require("user-groups-roles");
+const userRole = require("../roles/user.roles.js");
+
+const decodeAuthorization = authorization => decode(authorization);
 
 exports.addUser = async (req, res) => {
     try {    
-        const { Passport } = req.body;
-        let { Password } = req.body;
-        
-        const candidate = await User.findOne({
-            where: { Passport }
-        });
+        const decodedToken = await decodeAuthorization(req.headers.authorization)
+        const val = roles.getRoleRoutePrivilegeValue(decodedToken.role, "/add", "POST");
 
-        if(candidate) {
-            return res.status(406).json( { message: "Такой пациент уже существует" } );
-        } 
-
-        const hashedPassword = await bcrypt.hash(Password, 12);
-        
-        req.body.Password = hashedPassword;
-
-        await User.create(req.body);
-
-        return res.sendStatus(201).json({ message: "Пациент успешно зарегистрирован" });
+        if(val) {
+            const { Passport } = req.body;
+            let { Password } = req.body;
+            
+            const candidate = await User.findOne({
+                where: { Passport }
+            });
+    
+            if(candidate) {
+                return res.status(406).json( { message: "Такой пациент уже существует" } );
+            } 
+    
+            const hashedPassword = await bcrypt.hash(Password, 12);
+            
+            req.body.Password = hashedPassword;
+    
+            await User.create(req.body);
+    
+            return res.sendStatus(201).json({ message: "Пациент успешно зарегистрирован" });
+        } else {
+            return res.json({ error: "Запрещено" })
+        }
 
     } catch(err) {
         console.error(err);
-        return res.status(500).json({ message: "Что-то пошло не так, попробуйте снова" });
+        return res.status(500).json({ messageError: "Что-то пошло не так, попробуйте снова" });
     }
 };
 
 exports.getUsers = async (req, res) => {
     try {
-        await User.findAll()
+        const decodedToken = decodeAuthorization(req.headers.authorization)
+        const val = roles.getRoleRoutePrivilegeValue(decodedToken.role, "/get", "GET");
+
+        if(val) {
+            User.findAll()
             .then(users => res.send(users).json())
             .catch(err => res.sendStatus(500).json())
+        } else {
+            res.status(403).json({ error: "Запрещено" })
+        }
+        
 
     } catch(err) {
         console.error("ERROR: ", err);
-        res.send(404).json({ message: "Не найдено" })
+        res.send(404).json({ error: "Не найдено" })
     }    
 };
 
 exports.devDeleteUser = async (req, res) => {
     try {
-        await User.destroy({
-            where: {
-                idpacient: req.body.id
-            }
-        });
-        res.json({ message: "Удалено" })
+        const decodedToken = decodeAuthorization(req.headers.authorization)
+        const val = roles.getRoleRoutePrivilegeValue(decodedToken.role, "/delete", "POST");
+
+        if(val) {
+            await User.destroy({
+                where: {
+                    idpacient: req.body.id
+                }
+            });
+
+            res.status(200).json({ message: "Удалено" })
+        } else {
+            res.status(403).json({ error: "Запрещено" })
+        }
+        
     } catch (err) {
         console.error("ERROR :", err);
-        res.send(500).json({ message: "Что-то пошло не так, попробуйте снова" })
+        res.send(500).json({ error: "Что-то пошло не так, попробуйте снова" })
     }
 }
 
@@ -69,24 +97,21 @@ exports.authUser = async (req, res) => {
         });
     
         if(!user) {
-            return res.status(404).json({ message: "Такого пользователя не существует" });
+            return res.status(404).json({ error: "Такого пользователя не существует" });
         }
 
         const pass = await bcrypt.compare(Password, user.Password);
 
         if(!pass) {
-            return res.status(403).json({ message: "Неверный пароль" });
+            return res.status(403).json({ error: "Неверный пароль" });
         }
-
-        const userRules = rules(user); 
 
         const payload = {
             id: user.idpacient,
             name: user.Name,
             middlename: user.MiddleName,
             surname: user.Surname,
-            role: user.Role,
-            rules: userRules
+            role: user.Role
         }
         
 
@@ -95,7 +120,7 @@ exports.authUser = async (req, res) => {
         res.json({ token });
 
     } catch(err) {
-        res.status(500).json({ message: "Что-то пошло не так, попробуйте снова" });
+        res.status(500).json({ error: "Что-то пошло не так, попробуйте снова" });
     }
     
 
@@ -105,16 +130,23 @@ exports.authUser = async (req, res) => {
 exports.updateUserData = async (req, res) => {
     console.log(req.body);
     try {
-        await User.update(req.body, {
-            where: {
-                idpacient: req.body.id
-            }
-        })
+        const decodedToken = decodeAuthorization(req.headers.authorization)
+        const val = roles.getRoleRoutePrivilegeValue(decodedToken.role, "/update", "POST");
 
-        res.status(200).json({ message: "Обновлено" })
-        
+        if(val) {
+            await User.update(req.body, {
+                where: {
+                    idpacient: req.body.id
+                }
+            })
+            
+            res.status(200).json({ message: "Обновлено" })
+        } else {
+            res.status(403).json({ error: "Нет Доступа" })
+        }
+
     } catch(err) {
-        res.status(500).json({ message: "Что-то пошло не так, попробуйте снова" });
+        res.status(500).json({ error: "Что-то пошло не так, попробуйте снова" });
     }
 }
